Move test db into server dir and clean up database dir
