+++ server/src/db/connect.js
@@ -1,0 +1,9 @@
+const { Pool } = require('pg');
+
+const db = new Pool({
+  connectionString: process.env.DB_URL,
+});
+
+console.log('DB connection established: ', process.env.DB_URL);
+
+module.exports = db;
